Fix time_t prints for debian 13 armv7l
In Debian 13 armv7l, time_t and suseconds_t are "long long int" instead
of "long unsigned int" as seen on x86_64 and aarch64. Fix it by casting
to "long unsigned int".

Related: OS#6828
(cherry picked from commit 7dd8ca5e4b4b82498af6981af4a0fcad8980df30)

+++ tests/osmo-pcap-test/pcap.c
@@ -181,7 +181,7 @@ retry:
 		timersub(&elapsed_pcap, &elapsed_sys, &res);
 	}
 	printf("next packet comes in %lu.%.6lu seconds\n",
-		res.tv_sec, res.tv_usec);
+	       (unsigned int long) res.tv_sec, (unsigned int long) res.tv_usec);
 	osmo_timer_schedule(&p->timer, res.tv_sec, res.tv_usec);
 
 	return 0;
